+++ spring-cloud-deployer-spi-scheduler-test-app/Dockerfile
@@ -1,0 +1,7 @@
+FROM java:8-alpine
+
+ARG JAR_FILE
+
+ADD target/${JAR_FILE} spring-cloud-deployer-spi-scheduler-test-app.jar
+
+ENTRYPOINT ["java", "-jar", "/spring-cloud-deployer-spi-scheduler-test-app.jar"]
