+++ spring-cloud-deployer-spi-scheduler-test-app/pom.xml
@@ -6,9 +6,9 @@
 		<artifactId>spring-cloud-deployer-parent</artifactId>
 		<version>2.1.0.BUILD-SNAPSHOT</version>
 	</parent>
-	<artifactId>spring-cloud-deployer-scheduler-spi-test-app</artifactId>
+	<artifactId>spring-cloud-deployer-spi-scheduler-test-app</artifactId>
 	<properties>
-		<start-class>org.springframework.cloud.deployer.scheduler.spi.test.app.SchedulerIntegrationTestApplication</start-class>
+		<start-class>org.springframework.cloud.deployer.spi.scheduler.test.app.SchedulerIntegrationTestApplication</start-class>
 		<java.version>1.8</java.version>
 		<docker.image.prefix>springcloud</docker.image.prefix>
 		<dockerfile-maven-plugin.version>1.3.6</dockerfile-maven-plugin.version>
