+++ spring-cloud-deployer-spi-scheduler-test-app/src/main/java/org/springframework/cloud/deployer/spi/scheduler/test/app/SchedulerIntegrationTest.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.springframework.cloud.deployer.scheduler.spi.test.app;
+package org.springframework.cloud.deployer.spi.scheduler.test.app;
 
 import javax.annotation.PostConstruct;
 
@@ -25,8 +25,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.util.Assert;
-
-import static org.springframework.cloud.deployer.scheduler.spi.test.app.SchedulerIntegrationTestProperties.FUNNY_CHARACTERS;
 
 /**
  * An app that can misbehave, useful for integration testing of app deployers.
@@ -46,13 +44,13 @@ public class SchedulerIntegrationTest {
 	@PostConstruct
 	public void init() throws InterruptedException {
 		String parameterThatMayNeedEscaping = properties.getParameterThatMayNeedEscaping();
-		if (parameterThatMayNeedEscaping != null && !FUNNY_CHARACTERS.equals(parameterThatMayNeedEscaping)) {
-			throw new IllegalArgumentException(String.format("Expected 'parameterThatMayNeedEscaping' value to be equal to '%s', but was '%s'", FUNNY_CHARACTERS, parameterThatMayNeedEscaping));
+		if (parameterThatMayNeedEscaping != null && !SchedulerIntegrationTestProperties.FUNNY_CHARACTERS.equals(parameterThatMayNeedEscaping)) {
+			throw new IllegalArgumentException(String.format("Expected 'parameterThatMayNeedEscaping' value to be equal to '%s', but was '%s'", SchedulerIntegrationTestProperties.FUNNY_CHARACTERS, parameterThatMayNeedEscaping));
 		}
 
 		String commandLineArgValueThatMayNeedEscaping = properties.getCommandLineArgValueThatMayNeedEscaping();
-		if (commandLineArgValueThatMayNeedEscaping != null && !FUNNY_CHARACTERS.equals(commandLineArgValueThatMayNeedEscaping)) {
-			throw new IllegalArgumentException(String.format("Expected 'commandLineArgValueThatMayNeedEscaping' value to be equal to '%s', but was '%s'", FUNNY_CHARACTERS, commandLineArgValueThatMayNeedEscaping));
+		if (commandLineArgValueThatMayNeedEscaping != null && !SchedulerIntegrationTestProperties.FUNNY_CHARACTERS.equals(commandLineArgValueThatMayNeedEscaping)) {
+			throw new IllegalArgumentException(String.format("Expected 'commandLineArgValueThatMayNeedEscaping' value to be equal to '%s', but was '%s'", SchedulerIntegrationTestProperties.FUNNY_CHARACTERS, commandLineArgValueThatMayNeedEscaping));
 		}
 
 		Assert.notNull(properties.getInstanceIndex(), "instanceIndex should have been set by deployer or runtime");
