+++ spring-cloud-deployer-spi-scheduler-test-app/src/main/resources/application.properties
