+++ spring-cloud-deployer-spi/src/main/java/org/springframework/cloud/deployer/spi/scheduler/ScheduleInfo.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.springframework.cloud.deployer.scheduler.spi.core;
+package org.springframework.cloud.deployer.spi.scheduler;
 
 import java.util.Map;
